% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ParetoPosStable.R
\docType{package}
\name{ParetoPosStable}
\alias{ParetoPosStable}
\alias{ParetoPosStable-package}
\title{Computing, Fitting and Validating the PPS Distribution}
\description{
Statistical functions to describe a Pareto Positive Stable (PPS)
distribution and fit it to real data. Graphical and statistical tools to
validate the fits are included.
}
\details{
The main function you're likely to need from \pkg{ParetoPosStable} is
\code{\link{PPS.fit}}, in order to obtain a PPS fit from data. Validation can be obtain with \code{\link{GoF}} and \code{\link{plot}}.
}
\references{
Sarabia, J.M and Prieto, F. (2009). The Pareto-positive stable
distribution: A new descriptive model for city size data, \emph{Physica A:
Statistical Mechanics and its Applications}, \bold{388}(19), 4179-4191.
}

