% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/coef.PPSfit.R
\name{coef.PPSfit}
\alias{coef.PPSfit}
\title{Parameter estimates in a PPSfit Object}
\usage{
\method{coef}{PPSfit}(object, ...)
}
\arguments{
\item{object}{A \code{PPSfit} object, typically from \code{PPS.fit()}.}

\item{\dots}{Other arguments.}
}
\value{
A list with the parameter estimates.
}
\description{
\code{coef} returns the parameter estimates in a \code{PPSfit} Object
}
\examples{
x <- rPPS(50, 1.2, 100, 2.3)
fit <- PPS.fit(x)
coef(fit)
}
\seealso{
\code{\link{PPS.fit}}.
}

