% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/logLik.PPSfit.R
\name{logLik.PPSfit}
\alias{logLik.PPSfit}
\title{Log-likelihood value of a PPSfit Object}
\usage{
\method{logLik}{PPSfit}(object, ...)
}
\arguments{
\item{object}{A \code{PPSfit} Object.}

\item{\dots}{Other arguments.}
}
\value{
The log-likelihood.
}
\description{
It returns the log-likelihood value of a \code{PPSfit} Object
}
\examples{
x <- rPPS(50, 1.2, 100, 2.3)
fit <- PPS.fit(x)
logLik(fit)
}
\references{
Sarabia, J.M and Prieto, F. (2009). The Pareto-positive stable distribution: A new descriptive model for city size data, \emph{Physica A: Statistical Mechanics and its Applications}, \bold{388}(19), 4179-4191.
}
\seealso{
\code{\link{PPS.fit}}
}

